refactor(inference): split lib.rs into focused modules (TD-032)
Behavior-preserving split of parish-inference/src/lib.rs (~2390 LOC)
into five focused submodules:

- queue.rs   — InferenceRequest/Response, QueueRequest, InferencePriority, InferenceQueue
- logs.rs    — InferenceLogEntry, BoundedInferenceLog, InferenceLog, helpers
- any_client.rs — AnyClient, InferenceClients, build_client, StreamStats, TOKEN_CHANNEL_CAPACITY
- worker.rs  — InferenceWorkerConfig, spawn_inference_worker, inference_with_timeout
- timeout.rs — InferenceAwaitOutcome, await_inference_response, submit_json, submit_json_streaming

lib.rs is now a thin crate root: module declarations + re-exports only.
Public API (all parish_inference:: paths) is unchanged. Tests moved
alongside their logic. Part of #1200.
